Fix bounds checker frame increment
I forgot to increment the frame! This checked the same frame 20 times and for *most* movies the frame 1/20th of the way in worked fine. Tested the script with Kill Bill vol. 1 and the trimmer spit out a crazy small area, which turned out to be one of the opening credit titles. Now it correctly reads through all 20 frames spread throughout the video.