Buggy fix for buggy trim function
I have no idea why this works. The 'trim' option was always slow for no clear reason, and after changing to 10 sample frames for the trimmer I was getting IO errors that the program was trying to read past the end of the video file. From a little bit of research, apparently this is a bug in the OpenCV library due to some issue with h.264 encoding and b-frames when using the CAP_PROP_FRAME_COUNT 'set' option. Calling the 'read' function from the start of the video, like the visualizer function does, makes the selected sample frames match the visualizer frames (if n=10) and, for whatever reason, massively increases the processing speed. How bizarre...